feat: mode - reorganize
