+++ src/FlexLabs.EntityFrameworkCore.Upsert/Internal/Expressions/IKnownValue.cs
@@ -1,16 +1,15 @@
 ﻿using System.Collections.Generic;
 
-namespace FlexLabs.EntityFrameworkCore.Upsert.Internal
-{
+
+namespace FlexLabs.EntityFrameworkCore.Upsert.Internal.Expressions {
     /// <summary>
     /// Value of an expression
     /// </summary>
-    public interface IKnownValue
-    {
+    public interface IKnownValue {
         /// <summary>
         /// Get all the constants that are part of this value
         /// </summary>
-        /// <returns>A set of contsants used in this value</returns>
+        /// <returns>A set of constants used in this value</returns>
         IEnumerable<ConstantValue> GetConstantValues();
 
         /// <summary>
