+++ src/FlexLabs.EntityFrameworkCore.Upsert/Internal/Expressions/PropertyValue.cs
@@ -3,13 +3,11 @@ using System.Collections.Generic;
 using System.Linq.Expressions;
 
 
-namespace FlexLabs.EntityFrameworkCore.Upsert.Internal
-{
+namespace FlexLabs.EntityFrameworkCore.Upsert.Internal.Expressions {
     /// <summary>
     /// This class represents access to a property within an expression
     /// </summary>
-    public class PropertyValue : Expression, IKnownValue
-    {
+    public class PropertyValue : Expression, IKnownValue {
         /// <inheritdoc />
         public override ExpressionType NodeType => ExpressionType.Constant;
         /// <inheritdoc />
