+++ test/FlexLabs.EntityFrameworkCore.Upsert.Tests/Internal/ExpressionAssertionExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Linq.Expressions;
-using FlexLabs.EntityFrameworkCore.Upsert.Internal;
+using FlexLabs.EntityFrameworkCore.Upsert.Internal.Expressions;
 using FluentAssertions;
 using FluentAssertions.Execution;
 using FluentAssertions.Primitives;
